HOP-4213 : Create a new configuration perspective (more tabs)
