+++ plugins/misc/projects/src/main/resources/org/apache/hop/projects/config/messages/messages_en_US.properties
@@ -22,7 +22,7 @@ ProjectConfig.EnvironmentMandatory.Message=Use of an environment is mandatory
 ProjectConfig.DefaultProject.Message=The default project to use when none is specified
 ProjectConfig.DefaultEnvironment.Message=The default environment to use when none is specified
 ProjectConfig.ParentProject.Message=The parent project to propose when creating projects
-ProjectConfig.StdProjectFolder.Message=GUI: The standard projects folder proposed when creating projects
+ProjectConfig.StdProjectFolder.Message=The standard projects folder for new projects
 ProjectConfig.ProjectNotExists.Error.Header=Project doesn''t exists!
 ProjectConfig.ProjectNotExists.DefaultProject.Error.Message=Project ''{0}'' cannot be set as suggested default project because it does not exists!
 ProjectConfig.ProjectNotExists.StandardProject.Error.Message=Project ''{0}'' cannot be set as suggested parent project because it does not exists!
