+++ ui/src/main/resources/org/apache/hop/ui/core/dialog/messages/messages_en_US.properties
@@ -96,6 +96,7 @@ EnterOptionsDialog.DarkMode.Label=Dark mode?
 EnterOptionsDialog.GlobalZoom.Label=UI zoom level
 EnterOptionsDialog.DefaultLocale.Label=Preferred Language
 EnterOptionsDialog.DefaultPreviewSize.Label=Preview data batch size
+EnterOptionsDialog.WhatIsHopConfigSize.Label=This location can be changed with environment variable HOP_CONFIG
 EnterOptionsDialog.DialogMiddlePercentage.Label=Dialog middle percentage
 EnterOptionsDialog.EnableAutoCollapseCoreObjectTree.Label=Auto collapse palette tree
 EnterOptionsDialog.DefaultFont.Label=Default font
